add user subscribe

+++ app/models/user/UserCollect.scala
@@ -9,14 +9,14 @@ import java.sql.Timestamp
 case class UserCollect (
                     id: Option[Long],
                     uid:Long,
-                    typeId:Long,
+                    typeId:Int,
                     collectId:Long,
                     addTime:Option[Timestamp]
                     )
 class UserCollects(tag:Tag) extends Table[UserCollect](tag,"user_collect") {
   def id = column[Long]("id", O.PrimaryKey, O.AutoInc)
   def uid = column[Long]("uid")
-  def typeId = column[Long]("type_id")
+  def typeId = column[Int]("type_id")
   def collectId = column[Long]("collect_id")
   def addTime = column[Timestamp]("add_time")
   // Every table needs a * projection with the same type as the table's type parameter
